Change distance type to float in UltrasonicSensor
Updated the 'distance' member variable from int to float in the UltrasonicSensor class for more precise distance calculations. Also updated the calculation to use float literals.

+++ MainCode/main/classes.h
@@ -30,7 +30,7 @@ class UltrasonicSensor{
   int trigPin;
   int echoPin;
   long duration;
-  int distance;
+  float distance;
   
   public:
     UltrasonicSensor(int trigpin, int echopin){
@@ -48,7 +48,7 @@ class UltrasonicSensor{
       delayMicroseconds(10);
       digitalWrite(trigPin, LOW);
       duration  = pulseIn(echoPin, HIGH);
-      distance= duration*0.034/2;
+      distance= duration*0.034f/2.0f;
       return distance;
     }
 };
